List dishes by restaurant

+++ DAL/DishDB.cs
@@ -129,7 +129,7 @@ namespace DAL
                             dish.NAMEDISH = (string)dr["NAMEDISH"];
 
                             if (dr["PRICEDISH"] != null)
-                                dish.PRICEDISH = (int)dr["PRICEDISH"];
+                                dish.PRICEDISH = (double)dr["PRICEDISH"];
                         }
                     }
                 }
